% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EOBS_UK}
\alias{EOBS_UK}
\title{EOBS February Precipitation over the UK.}
\format{
A data frame with 71 rows and 2 variables:
\describe{
  \item{time}{The year, stored as yyyy-02-28 (or 29,we only have February averages).}
  \item{rr}{The rainfall rate, i.e. the average February precipitation (mm/d).}
  ...
}
}
\source{
\url{https://surfobs.climate.copernicus.eu/dataaccess/access_eobs.php}

\url{https://github.com/esowc/UNSEEN-open}
}
\usage{
EOBS_UK
}
\description{
A dataset containing February average precipitation (mm/d) over 1950-2020, downloaded from \url{https://surfobs.climate.copernicus.eu/dataaccess/access_eobs.php}
and pre-processed as part of the ECMWF Summer of Weather Code program: \url{https://github.com/esowc/UNSEEN-open}.
}
\keyword{datasets}
