% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_test.R
\name{Model_stability_boot}
\alias{Model_stability_boot}
\title{Test the model stability: 2. empirical extreme value plot}
\source{
Evaluation explaned in more detail in Kelder et al. 2020

Colorblind friendly palette  http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette
}
\usage{
Model_stability_boot(
  ensemble,
  var_name = "tprate",
  ld_name = "leadtime",
  lab = var_name,
  fontsize = 11
)
}
\arguments{
\item{ensemble}{The UNSEEN ensemble. This function expects an dataframe with variables leadtime, precipitation.}

\item{var_name}{The column name containing the variable to be analyzed. Defaults to "tprate".}

\item{ld_name}{The column name containing the leadtimes. Defaults to "leadtime".}

\item{lab}{The y-label. Defaults to the variable name (var_name).}

\item{fontsize}{The font size. Defaults to 11.}
}
\value{
a plot with the empirical return values of the pooled ensemble including confidence intervals.
Individual lead times are plotted on top.
}
\description{
We plot the density distribution of the different leadtimes using ggplot.
We want to show the confidence interval of the distribution of all lead times pooled together,
and test whether the individual lead times fall within these confidence intervals.
Therefore we bootstrap the pooled leadtimes into series with equal length of the individual leadtimes (875), with n=10000.
}
