% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_test.R
\name{Model_stability_density}
\alias{Model_stability_density}
\title{Test the model stability: 1. density plot}
\source{
Evaluation explained in more detail in Kelder et al. 2020

Colorblind friendly palette  http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette
}
\usage{
Model_stability_density(
  ensemble,
  var_name = "tprate",
  ld_name = "leadtime",
  lab = var_name
)
}
\arguments{
\item{ensemble}{The UNSEEN ensemble. This function expects an dataframe with the columns variable (precipitation) and leadtime.}

\item{var_name}{The column name containing the variable to be analyzed. Defaults to "tprate".}

\item{ld_name}{The column name containing the leadtimes. Defaults to "leadtime".}

\item{lab}{The x-label. Defaults to the variable name (var_name).}
}
\value{
a plot showing the empirical probability density distribution for each leadtime
}
\description{
We plot the density distribution of the different leadtimes using ggplot.
}
