% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_plot.R
\name{unseen_timeseries}
\alias{unseen_timeseries}
\title{Illustrate the timeseries of the UNSEEN ensemble and the observed using ggplot}
\usage{
unseen_timeseries(
  ensemble,
  obs,
  ensemble_yname = "tprate",
  ensemble_xname = "time",
  obs_yname = "rr",
  obs_xname = "time",
  ylab = "",
  title = ""
)
}
\arguments{
\item{ensemble}{The UNSEEN ensemble. This must be a dataframe containing at least the columns time and precipitation}

\item{obs}{The observations. This must be a dataframe with the variables year and precipitation}

\item{ensemble_yname}{The column name of the y variable. Defaults to "tprate"}

\item{ensemble_xname}{The column name of the x variable. Defaults to "time"}

\item{obs_yname}{The column name of the y variable. Defaults to "rr"}

\item{obs_xname}{The column name of the x variable. Defaults to "time"}

\item{ylab}{The y label (string)}

\item{title}{The title (string)}
}
\value{
a ggplot timeseries with x = year and y = variable
}
\description{
Illustrate the timeseries of the UNSEEN ensemble and the observed using ggplot
}
