% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_test.R
\name{stability_test}
\alias{stability_test}
\title{Test the model stability: It plots the output of 1. Model_stability_density(): a density plot and 2. Model_stability_boot(): an empirical extreme value plot}
\source{
Evaluation explaned in more detail in Kelder et al. 2020

Colorblind friendly palette  http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette
}
\usage{
stability_test(
  ensemble,
  var_name = "tprate",
  ld_name = "leadtime",
  lab = var_name,
  fontsize = 11,
  panel_labels = c("a", "b")
)
}
\arguments{
\item{ensemble}{The UNSEEN ensemble. This function expects an dataframe with variables leadtime, precipitation.}

\item{var_name}{The column name containing the variable to be analyzed. Defaults to "tprate".}

\item{ld_name}{The column name containing the leadtimes. Defaults to "leadtime".}

\item{lab}{The variable name plotted as x-label (first panel) and y-label (second panel). Defaults to the variable name (var_name).}

\item{fontsize}{The font size. Defaults to 11.}

\item{panel_labels}{The panel labels. Defaults to c("a", "b").}
}
\value{
a plot with the empirical return values of the pooled ensemble including confidence intervals.
Individual lead times are plotted on top.
}
\description{
Test the model stability: It plots the output of 1. Model_stability_density(): a density plot and 2. Model_stability_boot(): an empirical extreme value plot
}
\seealso{
[Model_stability_density()] [Model_stability_boot()]d
}
